// src/visual-dashboard-frontend/src/interfaces/schemasReducer.ts
export interface TabInList {
    id: number
    name: string,
    sensors_ids: any[]
}

export interface SchemaDetails {
    comment: string
    created_dt: string
    id: number
    is_published: boolean
    modified_by: number
    name: string
    schema_tabs: TabInList[]
    update_dt: string
    device_ids: number[]
}